app/requirements.txt
fastapi
uvicorn
pandas
mysql-connector-python
python-multipart